requirements.txt
asyncio
websockets
requests
numpy
sounddevice
python-dotenv